Basic page allocation logic
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@2874 42af7a65-404d-4744-a932-0658087f49c3

+++ arch/arm/include/arm/irq.h
@@ -164,7 +164,7 @@ struct xcptcontext
    */
 
 #ifdef CONFIG_PAGING
-  uint32_t far;
+  uintptr_t far;
 #endif
 };
 #endif
